Support patch for publish.
Cargo doesn't support patch for publishing. I don't want to claim this
is a bug, but it really block some features. User must wait the
dependency to release a new version otherwise, it can't publish a
package with the patch of the dependency.

This change will support patch during the publishing, but I don't
suggest you use this hack version to publish a package to crates.io.
You should only use this hack version to publish to your private
repository.